refactoring and further work on classes and exposure rating
